==> jupyter_notebooks/06 - Modelling and Evaluation - Cluster Sklearn.ipynb <==
# **Cluster**

## Objectives

* Fit and evaluate a cluster model to group similar data
* Analyse the clusters against the diagnostic (malignant or benign)
* Understand the profile for each cluster

## Inputs

* Outputs/datasets/collection/breast-cancer.csv
* Instructions on which variables to use for data cleaning and feature engineering, which are found in their respective notebooks.

## Outputs

* Cluster Pipeline
* Train Set
* Most important features to define a cluster plot
* Clusters Profile Description
* Cluster Silhouette

## Additional Comments

* This notebook was written based on the guidelines provided in the walkthrough project 2: 'Churnometer'.
* This notebook relates to the Data Understanding step of Crisp-DM methodology.
* This notebook and the following will represent the learning outcome after following the Code Institute - Predictive Analytics and Machine Learning module.


---

# Change working directory

* We are assuming you will store the notebooks in a subfolder, therefore when running the notebook in the editor, you will need to change the working directory

We need to change the working directory from its current folder to its parent folder
* We access the current directory with os.getcwd()

import os
current_dir = os.getcwd() 
current_dir

We want to make the parent of the current directory the new current directory
* os.path.dirname() gets the parent directory
* os.chir() defines the new current directory

os.chdir(os.path.dirname(current_dir))
print("You set a new current directory")

Confirm the new current directory

current_dir = os.getcwd()
current_dir

# Load Data

import pandas as pd
df = (pd.read_csv("outputs/datasets/collection/breast-cancer.csv")
    .drop(['id', 'diagnosis'], axis=1)
    )
print(df.shape)
df.head(3)

# Cluster Pipeline with all data

## ML Pipeline for Cluster

* Our objective is to cluster similar data points and then analyse the clusters against the diagnostic (malignant or benign).

* As a result, we will use only the thirty features (all variables but Diagnostic) to fit the cluster pipeline.

from sklearn.pipeline import Pipeline

### Feat Scaling
from sklearn.preprocessing import StandardScaler

### PCA
from sklearn.decomposition import PCA

### ML algorithm
from sklearn.cluster import KMeans

def PipelineCluster():
    pipeline_base = Pipeline([
        
        ("scaler", StandardScaler()),
        ("PCA", PCA(n_components=50, random_state=0)),
        ("model", KMeans(n_clusters=50, random_state=0)),
    ])

    return pipeline_base

PipelineCluster()

## Principal Component Analysis (PCA)

* We are interested to find the most suitable n_components, then we update the value in the ML Pipeline for Cluster

* To reach that, we will create an object based on PipelineCluster(), then remove the last two steps (PCA and model): .steps[:-2]

* Finally, the pipeline_pca scales the data, so we can apply PCA afterwards

pipeline_cluster = PipelineCluster()
pipeline_pca = Pipeline(pipeline_cluster.steps[:-2])
df_pca = pipeline_pca.fit_transform(df)

print(df_pca.shape,'\n', type(df_pca))

* Next, we apply PCA separately to the scaled data

%matplotlib inline
# This line is used to display matplotlib plots inline within the Jupyter Notebook.

import numpy as np
import matplotlib.pyplot as plt
import seaborn as sns
sns.set_style("whitegrid")

n_components = 16


def pca_components_analysis(df_pca, n_components):
    pca = PCA(n_components=n_components).fit(df_pca)
    x_PCA = pca.transform(df_pca)  # array with transformed PCA

    ComponentsList = ["Component " + str(number)
                    for number in range(n_components)]
    dfExplVarRatio = pd.DataFrame(
        data=np.round(100 * pca.explained_variance_ratio_, 3),
        index=ComponentsList,
        columns=['Explained Variance Ratio (%)'])

    dfExplVarRatio['Accumulated Variance'] = dfExplVarRatio['Explained Variance Ratio (%)'].cumsum(
    )

    PercentageOfDataExplained = dfExplVarRatio['Explained Variance Ratio (%)'].sum(
    )

    print(
        f"* The {n_components} components explain {round(PercentageOfDataExplained,2)}% of the data \n")
    plt.figure(figsize=(9, 6))
    sns.lineplot(data=dfExplVarRatio,  marker="o")
    plt.xticks(rotation=90)
    plt.yticks(np.arange(0, 110, 10))
    plt.show()


pca_components_analysis(df_pca=df_pca, n_components=n_components)

* With seven components we can achieve a bit more than 90% of data variance judging by the above figure

n_components = 7

pca = PCA(n_components=n_components).fit(df_pca)
x_PCA = pca.transform(df_pca)

ComponentsList = ["Component " + str(number) for number in range(n_components)]
dfExplVarRatio = pd.DataFrame(
    data= np.round(100 * pca.explained_variance_ratio_ ,3),
    index=ComponentsList,
    columns=['Explained Variance Ratio (%)'])

PercentageOfDataExplained = dfExplVarRatio['Explained Variance Ratio (%)'].sum()

print(f"* The {n_components} components explain {round(PercentageOfDataExplained,2)}% of the data \n")
print(dfExplVarRatio)

* We rewrite the PipelineCluster(), updating n_components to 7

def PipelineCluster():
    pipeline_base = Pipeline([

        ("scaler", StandardScaler()),
        ("PCA", PCA(n_components=7, random_state=0)),
        ("model", KMeans(n_clusters=50, random_state=0)),
    ])

    return pipeline_base

PipelineCluster()

## Elbow Method and Silhouette Score

* We will combine 2 techniques (Elbow Method and Silhouette Score) to find the optimal value for the number of clusters

* We will transform the data up to the point that it will hit the model, for Elbow Method and Silhouette score.
  * Therefore we remove the last step (.steps[:-1]) and fit_transform pipeline_analysis to the data.

pipeline_cluster = PipelineCluster()
pipeline_analysis = Pipeline(pipeline_cluster.steps[:-1])
df_analysis = pipeline_analysis.fit_transform(df)

print(df_analysis.shape,'\n', type(df_analysis))

### Elbow Method

* We use KElbowVisualizer() from YellowbrickElbow Analysis to implement the Elbow Method

* We pass in as arguments the algorithm we want (KMeans) and the range for the number of clusters we want to try

from yellowbrick.cluster import KElbowVisualizer

plt.rcParams['font.family'] = 'DejaVu Sans'
plt.rcParams['font.sans-serif'] = ['DejaVu Sans']

visualizer = KElbowVisualizer(KMeans(random_state=0), k=(1, 11)) # 11 is not inclusive
visualizer.fit(df_analysis)
visualizer.show()

From the visualizer we can deduce -

* The plot suggests 3 clusters
* Between 2 and 5, the values have a sharp and steep falloff.|
* Outside this range, it does not fall off in a similar manner.

### Silhouette score

* By Silhouette score, we can interpret and validate the consistency within clusters, which is based on the mean intra-cluster distance and mean nearest-cluster distance for each data point.

The silhouette score range is from -1 to +1.

* “+1” means that a clustered data point is dense and properly separated from other clusters.
* A score close to 0 means the clustered data point is overlapping with another cluster.
* A negative score means that the clustered data point may be wrong; it may even belong to another cluster.

from yellowbrick.cluster import SilhouetteVisualizer

plt.rcParams['font.family'] = 'DejaVu Sans'
plt.rcParams['font.sans-serif'] = ['DejaVu Sans']
print("=== Average Silhouette Score for different number of clusters ===")
visualizer = KElbowVisualizer(KMeans(random_state=0), k=(2,5), metric='silhouette')
visualizer.fit(df_analysis) 
visualizer.show() 
plt.show()
print("\n")

for n_clusters in np.arange(start=2,stop=11):

    print(f"=== Silhouette plot for {n_clusters} Clusters ===")
    visualizer = SilhouetteVisualizer(estimator = KMeans(n_clusters=n_clusters, random_state=0),
                                        colors = 'yellowbrick')
    visualizer.fit(df_analysis)
    visualizer.show()
    print("\n")

**Optimal Value for clusters -**

* Elbow Method says 3.
* The average Silhouette Score says 2, but by the Silhouette Plot from 3 clusters, it is better to use than 2 clusters.
* As a result, we will pick 3, since the Elbow Method and Silhouette Plot both support that decision.

* **We rewrite the PipelineCluster(), updating n_cluster to 3**

def PipelineCluster():
    pipeline_base = Pipeline([

        ("scaler", StandardScaler()),
        ("PCA", PCA(n_components=7, random_state=0)),
        ("model", KMeans(n_clusters=3, random_state=0)),
    ])

    return pipeline_base

PipelineCluster()

## Fit Cluster Pipeline

* Quick recap of our data for training cluster pipeline

X = df.copy()
print(X.shape)
X.head(3)

* Fit Cluster pipeline

pipeline_cluster = PipelineCluster()
pipeline_cluster.fit(X)

## Add cluster predictions to dataset

* **We add a column "Clusters" (with the Cluster Pipeline predictions) to X**

X['Clusters'] = pipeline_cluster['model'].labels_
print(X.shape)
X.head(3)

* We are interested to know the cluster frequency

print(f"* Clusters frequencies \n{ X['Clusters'].value_counts(normalize=True).to_frame().round(2)} \n\n")
X['Clusters'].value_counts().sort_values().plot(kind='bar')
plt.show()

*  There are three clusters, and the counting starts from 0.

* We note that the algorithm found that the majority of the data (63%) belongs to cluster number 1.

* Cluster 0 and 2 contain 19% and 18% of the data.

import seaborn as sns
import matplotlib.pyplot as plt
sns.set_style("whitegrid")
plt.figure(figsize=(10, 6))
sns.scatterplot(x=df_analysis[:, 0], y=df_analysis[:, 1],
                hue=X['Clusters'], palette='Set1', alpha=0.6)
plt.scatter(x=pipeline_cluster['model'].cluster_centers_[:, 0], y=pipeline_cluster['model'].cluster_centers_[:, 1],
            marker="x", s=169, linewidths=3, color="black")
plt.xlabel("PCA Component 0")
plt.ylabel("PCA Component 1")
plt.title("PCA Components colored by Clusters")
plt.show()

From the scatterpolt above -

* Cluster 1 has the highest density of data.
* Overlap quantity among the data is very low which serves as a good point for cluster analysis.

* We save the cluster predictions from this pipeline to use in the future.

cluster_predictions_with_all_variables = X['Clusters']
cluster_predictions_with_all_variables

## Fit a classifier, where the target is cluster predictions and features remaining variables

* Our new dataset has Clusters as a variable.
* We use a technique where Clusters will be the target for a classifier, and the remaining variables will be features for that target.

* We copy X to a DataFrame df_clf

df_clf = X.copy()
print(df_clf.shape)
df_clf.tail(3)

* Split Train and Test sets

from sklearn.model_selection import train_test_split
X_train, X_test, y_train, y_test = train_test_split(
    df_clf.drop(['Clusters'], axis=1),
    df_clf['Clusters'],
    test_size=0.2,
    random_state=0
)

print(X_train.shape, y_train.shape, X_test.shape, y_test.shape)

* Create classifier pipeline steps.
* We are considering a model that typically offers good results, and feature's importance can be assessed with .features_importance_ using a tree-based algorithm. 
* We are using AdaBoostClassifier since it had good performance in the previous notebook.

# Feat Selection
from sklearn.feature_selection import SelectFromModel

# ML algorithm
from sklearn.ensemble import GradientBoostingClassifier

def PipelineClf2ExplainClusters():
    pipeline_base = Pipeline([
        ("scaler", StandardScaler()),
        ("feat_selection", SelectFromModel(GradientBoostingClassifier(random_state=0)) ),
        ("model", GradientBoostingClassifier(random_state=0)),
    ])

    return pipeline_base

PipelineClf2ExplainClusters()

* Fit the classifier to the training data

pipeline_clf_cluster = PipelineClf2ExplainClusters()
pipeline_clf_cluster.fit(X_train, y_train)

## Evaluate classifier performance on Train and Test Sets

from sklearn.metrics import classification_report
print(classification_report(y_train, pipeline_clf_cluster.predict(X_train)))

print(classification_report(y_test, pipeline_clf_cluster.predict(X_test)))

## Assess the most important Features that define a cluster

pipeline_clf_cluster

# Taking pipeline steps upto classifier without including it
pipeline_feat_select = Pipeline(pipeline_clf_cluster.steps[:-2])

# Fit-transform the feature selection pipeline
X_train_selected = pipeline_feat_select.fit_transform(X_train, y_train)

# The columns that survive the selection
selected_columns = X_train.columns[pipeline_clf_cluster['feat_selection'].get_support()]

# create DataFrame to display feature importance
df_feature_importance = (pd.DataFrame({
    'Feature': selected_columns,
    'Importance': pipeline_clf_cluster['model'].feature_importances_})
                         .sort_values(by='Importance', ascending=False)
)

# assign best features in importance order
best_features = df_feature_importance['Feature'].to_list()

# Most important features statement and plot
print(f"* These are the {len(best_features)} most important features in descending order. "
    f"The model was trained on them: \n{best_features} \n")
df_feature_importance.plot(kind='bar', x='Feature', y='Importance')
plt.show()

These are the 4 most important features in descending order -

* ['concavity_mean', 'perimeter_worst', 'fractal_dimension_worst', 'perimeter_mean']

best_features_pipeline_all_variables = best_features
best_features_pipeline_all_variables

## Cluster Analysis

* We load function that plots a table with description for all Clusters

def DescriptionAllClusters(df, decimal_points=3):

  DescriptionAllClusters = pd.DataFrame(columns=df.drop(['Clusters'],axis=1).columns)
  # iterate on each cluster , calls Clusters_IndividualDescription()
  for cluster in df.sort_values(by='Clusters')['Clusters'].unique():
    
      EDA_ClusterSubset = df.query(f"Clusters == {cluster}").drop(['Clusters'],axis=1)
      ClusterDescription = Clusters_IndividualDescription(EDA_ClusterSubset,cluster,decimal_points)
      DescriptionAllClusters = pd.concat([DescriptionAllClusters, ClusterDescription], ignore_index=True)

  
  DescriptionAllClusters.set_index(['Cluster'],inplace=True)
  return DescriptionAllClusters


def Clusters_IndividualDescription(EDA_Cluster,cluster, decimal_points):

  ClustersDescription = pd.DataFrame(columns=EDA_Cluster.columns)
  # for a given cluster, iterate in all columns
  # if the variable is numerical, calculate the IQR: display as Q1 -- Q3.
    # That will show the range for the most common values for the numerical variable
  # if the variable is categorical, count the frequencies and display the top 3 most frequent
    # That will show the most common levels for the category

  for col in EDA_Cluster.columns:
    
    try:  # eventually a given cluster will have only missing data for a given variable
      
      if EDA_Cluster[col].dtypes == 'object':
        
        top_frequencies = EDA_Cluster.dropna(subset=[col])[[col]].value_counts(normalize=True).nlargest(n=3)
        Description = ''
        
        for x in range(len(top_frequencies)):
          freq = top_frequencies.iloc[x]
          category = top_frequencies.index[x][0]
          CategoryPercentage = int(round(freq*100,0))
          statement =  f"'{category}': {CategoryPercentage}% , "  
          Description = Description + statement
        
        ClustersDescription.at[0,col] = Description[:-2]


      
      elif EDA_Cluster[col].dtypes in ['float', 'int']:
        DescStats = EDA_Cluster.dropna(subset=[col])[[col]].describe()
        Q1 = round(DescStats.iloc[4,0], decimal_points)
        Q3 = round(DescStats.iloc[6,0], decimal_points)
        Description = f"{Q1} -- {Q3}"
        ClustersDescription.at[0,col] = Description
    
    
    except Exception as e:
      ClustersDescription.at[0,col] = 'Not available'
      print(f"** Error Exception: {e} - cluster {cluster}, variable {col}")
  
  ClustersDescription['Cluster'] = str(cluster)
  
  return ClustersDescription


* We load a custom function to plot cluster distribution per Variable (absolute and relative levels)

import plotly.express as px


def cluster_distribution_per_variable(df, target):
    """
    The data should have 2 variables, the cluster predictions and
    the variable you want to analyze with, in this case we call "target".
    We use plotly express to create 2 plots:
    Cluster distribution across the target.
    Relative presence of the target level in each cluster.
    """
    df_bar_plot = df.groupby(['Clusters', target]).size().reset_index(name='Count')
    df_bar_plot.columns = ['Clusters', target, 'Count']
    df_bar_plot[target] = df_bar_plot[target].astype('object')

    print(f"Clusters distribution across {target} levels")
    fig = px.bar(df_bar_plot, x='Clusters', y='Count',
                 color=target, width=800, height=500)
    fig.update_layout(xaxis=dict(tickmode='array',
                      tickvals=df['Clusters'].unique()))
    fig.show(renderer='jupyterlab')

    df_relative = (df
                   .groupby(["Clusters", target])
                   .size()
                   .unstack(fill_value=0)
                   .apply(lambda x: 100 * x / x.sum(), axis=1)
                   .stack()
                   .reset_index(name='Relative Percentage (%)')
                   .sort_values(by=['Clusters', target])
                   )

    print(f"Relative Percentage (%) of {target} in each cluster")
    fig = px.line(df_relative, x='Clusters', y='Relative Percentage (%)',
                  color=target, width=800, height=500)
    fig.update_layout(xaxis=dict(tickmode='array',
                      tickvals=df['Clusters'].unique()))
    fig.update_traces(mode='markers+lines')
    fig.show(renderer='jupyterlab')

* We create a DataFrame that contains best features and Clusters Predictions since we want to analyse the patterns for each cluster

df_cluster_profile = df_clf.copy()
df_cluster_profile = df_cluster_profile.filter(items=best_features + ['Clusters'], axis=1)
print(df_cluster_profile.shape)
df_cluster_profile.tail(3)

* We want to analyse diagnosis levels also, so we load its data

df_diagnosis = pd.read_csv("outputs/datasets/collection/breast-cancer.csv").filter(['diagnosis'])
df_diagnosis['diagnosis'] = df_diagnosis['diagnosis'].astype('object')
df_diagnosis.head(3)

### Cluster profile based on the best features

pd.set_option('display.max_colwidth', None)
clusters_profile = df=pd.concat([df_cluster_profile,df_diagnosis], axis=1)
clusters_profile

pd.set_option('display.max_colwidth', None)
clusters_profile = DescriptionAllClusters(df=pd.concat([df_cluster_profile,df_diagnosis], axis=1),
                                          decimal_points=3)
clusters_profile

## Cluster analysis from each profile

Since we have set '0' as Benign (B) and '1' as Malignant (M), from the above profiling we can describe each clusters in the following -

* **Cluster 0 Profile**

    - In Cluster 0, mean concavity values range from 0.133 to 0.225, indicating moderately irregular tumor contours. The worst perimeter measurements, ranging from 145.7 to 174.175, suggest larger tumor sizes, while the worst fractal dimension falls between 0.076 and 0.093, indicating moderate structural complexity. Mean perimeter values (120.875 to 136.875) reflect less compact shapes. Notably, 100% of diagnoses in this cluster are malignant (M), aligning with high-risk features.

* **Cluster 1 Profile**

    - Cluster 1 shows mean concavity from 0.02 to 0.058, indicating smoother tumor margins. The worst perimeter is significantly smaller, ranging from 79.848 to 99.285, while the worst fractal dimension (0.069 to 0.082) suggests low structural disorder. The mean perimeter (72.185 to 88.345) reflects compact tumor shapes. Diagnoses in this cluster are predominantly benign (B: 90%), with only 10% malignant (M), consistent with a low-risk profile.

- **Cluster 2 Profile**

    - In Cluster 2, mean concavity ranges from 0.104 to 0.181, indicating moderately irregular shapes. The worst perimeter spans 94.22 to 122.4, suggesting mid-range tumor size, while the mean perimeter varies between 78.99 and 102.4. The worst fractal dimension (0.095 to 0.119) is the highest among all clusters, showing more chaotic tumor structures. Diagnoses are mixed, with 64% malignant (M) and 36% benign (B), suggesting potential intermediate-risk cases requiring closer evaluation.

### Clusters distribution across diagnosis levels & Relative Percentage of diagnosis in each cluster

df_cluster_vs_diagnosis = df_diagnosis.copy()
df_cluster_vs_diagnosis['Clusters'] = X['Clusters']
cluster_distribution_per_variable(df=df_cluster_vs_diagnosis, target='diagnosis')

## Fit New Cluster Pipeline with most important features

* In order to reduce feature space, we will study the trade-off between the previous Cluster Pipeline (fitted with all variables) and Pipeline using the variables that are most important to define the clusters from the previous pipeline

best_features_pipeline_all_variables

### Define trade-off and metrics to compare new and previous Cluster Pipeline

To evaluate this trade-off we will -

1. Conduct a elbow method and silhouette analysis and check if the same number of clusters is suggested
2. Fit new cluster pipeline and compare if the predictions from this pipeline are "equivalent" to the predictions from the previous pipeline
3. Fit a classifier to explain cluster, and check if performance on Train and Test sets is similar to the previous pipeline
4. Check if the most important features for the classifier are the same from the previous pipeline
5. Compare if the cluster profile from both pipelines are "equivalent"

* Should we express our willingness to support them, we may implement a cluster pipeline that utilizes the features most indicative of the clusters identified in the previous pipeline.

### Subset data with the most relevant variables

df_reduced = df.filter(best_features_pipeline_all_variables)
print(df_reduced.shape)
df_reduced.head(3)

### Rewrite Cluster Pipeline

def PipelineCluster():
    pipeline_base = Pipeline([
        ("scaler", StandardScaler()),
        ("model", KMeans(n_clusters=3, random_state=0)),
    ])

    return pipeline_base

PipelineCluster()

### Apply Elbow Method and Silhouette analysis

pipeline_cluster = PipelineCluster()
pipeline_analysis = Pipeline(pipeline_cluster.steps[:-1])
df_analysis = pipeline_analysis.fit_transform(df_reduced)

print(df_analysis.shape,'\n', type(df_analysis))

#### Elbow Analysis

from yellowbrick.cluster import KElbowVisualizer
print("=== Average Elbow Score for different number of clusters ===")
visualizer = KElbowVisualizer(KMeans(random_state=0), k=(1,11))
visualizer.fit(df_analysis) 
visualizer.show() 
plt.show()

From the visualizer we can deduce -

* The plot suggests 3 clusters
* Between 2 and 5, the values have a sharp and steep falloff.
* Outside this range, it does not fall off in a similar manner.

#### Silhoutte Score

from yellowbrick.cluster import SilhouetteVisualizer

n_cluster_start, n_cluster_stop = 2, 5

print("=== Average Silhouette Score for different number of clusters ===")
visualizer = KElbowVisualizer(KMeans(random_state=0), k=(
    n_cluster_start, n_cluster_stop), metric='silhouette')
visualizer.fit(df_analysis)
visualizer.show()
plt.show()
print("\n")


for n_clusters in np.arange(start=n_cluster_start, stop=n_cluster_stop):

    print(f"=== Silhouette plot for {n_clusters} Clusters ===")
    visualizer = SilhouetteVisualizer(estimator=KMeans(n_clusters=n_clusters, random_state=0),
                                      colors='yellowbrick')
    visualizer.fit(df_analysis)
    visualizer.show()
    plt.show()
    print("\n")

**Optimal Value for clusters with best features-**

* Elbow Method says 3.
* The average Silhouette Score says 2, but by the Silhouette Plot from 3 clusters, it is better to use than 2 clusters.
* As a result, we will pick 3, since the Elbow Method and Silhouette Plot both support that decision.

## Fit New Cluster Pipeline

* We set X as our training set for the cluster. It is a copy of df_reduced

X = df_reduced.copy()
print(X.shape)
X.head(3)

### Fit Cluster pipeline

pipeline_cluster = PipelineCluster()
pipeline_cluster.fit(X)

### Add cluster predictions to dataset

* We add a column "Clusters" (with the cluster pipeline predictions) to the dataset

X['Clusters'] = pipeline_cluster['model'].labels_
print(X.shape)
X.head(3)

print(f"* Clusters frequencies \n{ X['Clusters'].value_counts(normalize=True).to_frame().round(2)} \n\n")
X['Clusters'].value_counts().sort_values().plot(kind='bar')
plt.show()

## Compare current cluster predictions to previous cluster predictions

We just fitted a new cluster pipeline and want to compare if its predictions are "equivalent" to the previous cluster.

* These are the predictions from the previous cluster pipeline - trained with all variables

cluster_predictions_with_all_variables

* And these are the predictions from current cluster pipeline, trained with ['concavity_mean', 'area_worst', 'fractal_dimension_worst', 'perimeter_worst']

cluster_predictions_with_best_features = X['Clusters'] 
cluster_predictions_with_best_features

#### We use a confusion matrix to evaluate if the predictions of both pipelines are "equivalent"

from sklearn.metrics import confusion_matrix
print(confusion_matrix(cluster_predictions_with_all_variables, cluster_predictions_with_best_features))

* We can see that the cluster alignment is matched perfectly between the two and the data spread is similar, other than cluster 2 being heavier on the Malignant side.

#### Fit a classifier, where the target is cluster predictions and features remaining variables

df_clf = X.copy()
print(df_clf.shape)
df_clf.head(3)

* Split Train and Test sets

from sklearn.model_selection import train_test_split

X_train, X_test, y_train, y_test = train_test_split(
    df_clf.drop(['Clusters'], axis=1),
    df_clf['Clusters'],
    test_size=0.2,
    random_state=0
)

print(X_train.shape, X_test.shape)

* Rewrite pipeline to explain clusters

def PipelineClf2ExplainClusters():
    pipeline_base = Pipeline([
        ("scaler", StandardScaler()),
        # we don't consider feature selection step, since we know which features to consider
        ("model", GradientBoostingClassifier(random_state=0)),

    ])
    return pipeline_base


PipelineClf2ExplainClusters()

### Fit a classifier, where the target is cluster labels and features remaining variables

* We create and fit a classifier pipeline to learn the feature importance when defining a cluster

pipeline_clf_cluster = PipelineClf2ExplainClusters()
pipeline_clf_cluster.fit(X_train,y_train)

## Evaluate classifier performance on Train and Test Sets

print(classification_report(y_train, pipeline_clf_cluster.predict(X_train)))


print(classification_report(y_test, pipeline_clf_cluster.predict(X_test)))

## Assess Most Important Features

# since we don't have feature selection step in this pipeline, best_features is Xtrain columns
best_features = X_train.columns.to_list()

# create a DataFrame to display feature importance
df_feature_importance = (pd.DataFrame(data={
    'Feature': best_features,
    'Importance': pipeline_clf_cluster['model'].feature_importances_})
    .sort_values(by='Importance', ascending=False)
)

best_features = df_feature_importance['Feature'].to_list()

# Most important features statement and plot
print(f"* These are the {len(best_features)} most important features in descending order. "
      f"The model was trained on them: \n{df_feature_importance['Feature'].to_list()}")

df_feature_importance.plot(kind='bar', x='Feature', y='Importance')
plt.show()

## Cluster Analysis

* We create a DataFrame that contains the best features and Clusters Predictions: we want to analyse the patterns for each cluster.

df_cluster_profile = df_clf.copy()
df_cluster_profile = df_cluster_profile.filter(items=best_features + ['Clusters'], axis=1)
df_cluster_profile.head(3)

* We analyse diagnosis levels

df_diagnosis = pd.read_csv("outputs/datasets/collection/breast-cancer.csv").filter(['diagnosis'])
df_diagnosis['diagnosis'] = df_diagnosis['diagnosis'].astype('object')
df_diagnosis.head(3)

### Cluster profile on most important features

pd.set_option('display.max_colwidth', None)
clusters_profile = DescriptionAllClusters(df= pd.concat([df_cluster_profile,df_diagnosis], axis=1))
clusters_profile

### Cluster Profile Analysis

1. **Cluster 0 Profile**
* In Cluster 0, mean concavity values range from 0.112 to 0.198, indicating moderately irregular tumor contours. The worst perimeter measurements range from 97.692 to 119.7, suggesting mid-range tumor size, while the worst fractal dimension falls between 0.104 and 0.124, indicating high structural complexity. Mean perimeter values of 81.505 to 98.165 reflect moderately sized tumors. Notably, 69% of diagnoses in this cluster are malignant (M), with 31% benign (B), suggesting a significant risk of malignancy with some borderline cases.

2. **Cluster 1 Profile**
* Cluster 1 shows mean concavity from 0.02 to 0.059, indicating smooth tumor margins. The worst perimeter is significantly smaller, ranging from 79.7 to 98.385, while the worst fractal dimension (0.069 to 0.083) suggests low structural disorder. The mean perimeter ranges from 71.85 to 87.86, reflecting compact tumor shapes. Diagnoses in this cluster are predominantly benign (B: 92%), with only 8% malignant (M), consistent with a low-risk profile.

3. **Cluster 2 Profile**
* In Cluster 2, mean concavity ranges from 0.125 to 0.219, indicating highly irregular tumor shapes. The worst perimeter spans from 140.35 to 170.15, suggesting large tumor size, while the mean perimeter varies between 115.8 and 134.7. The worst fractal dimension (0.076 to 0.092) shows moderate structural complexity. Notably, 100% of the diagnoses in this cluster are malignant (M), indicating an aggressive, high-risk tumor profile that requires immediate attention.

### Clusters distribution across diagnosis levels & Relative Percentage of diagnosis in each cluster

df_cluster_vs_diagnosis = df_diagnosis.copy()
df_cluster_vs_diagnosis['Clusters'] = X['Clusters']
cluster_distribution_per_variable(df=df_cluster_vs_diagnosis, target='diagnosis')

### Comparative analysis of the cluster profiles between the original pipeline and the new pipeline

Here’s a concise comparative analysis of the cluster profiles between the original pipeline and the new pipeline:

* **Cluster 0 Comparison**

    - The original Cluster 0 (100% malignant) has split into two distinct groups in the new model. The high-risk profile (original 0) now appears as new Cluster 2 (100% malignant), while a new intermediate-risk Cluster 0 emerges (69% malignant). Key changes include: perimeter ranges shifting downward (145.7-174.175 to 97.692-119.7) and fractal dimension increasing (0.076-0.093 to 0.104-0.124), suggesting the model now better separates moderate-risk cases from aggressive tumors.

* **Cluster 1 Comparison**

    - The benign Cluster 1 remains stable between versions, maintaining: nearly identical concavity (0.02-0.058 to 0.02-0.059), perimeter ranges (79.848-99.285 to 79.7-98.385), and fractal dimension (0.069-0.082 to 0.069-0.083). The benign proportion improved slightly (90% to 92%), while malignant cases decreased (10% to 8%), indicating enhanced specificity for non-cancerous tumors.

* **Cluster 2 Comparison**

    - The original intermediate Cluster 2 (64% malignant) has been replaced by a pure malignant cluster in the new model (100%). This new Cluster 2 combines features from both original clusters: inheriting high concavity (0.125-0.219 vs original 0.133-0.225) and perimeter (140.35-170.15 vs original 145.7-174.175) from the original malignant group, while showing tighter fractal dimension control (0.076-0.092 vs original 0.095-0.119). This suggests the revised model more confidently classifies aggressive malignancies.

Overall, the four-feature model aligns well with the original 30-feature version, particularly for clear benign or malignant cases. Cluster 2's increased malignancy percentage suggests these features are effective in identifying high-risk tumors, while the broader ranges indicate slightly less precision in borderline cases. The consistency across Cluster 0 and Cluster 1 further confirms the reliability of the 4 feature-cluster pipeline.

---

# Conclusion and Pushing files to Repo

## Which pipeline should I deploy?

We recap the criteria we consider to evaluate the trade-off -

* *Conduct an elbow method and silhouette analysis and check if the same number of clusters is suggested.*
* *Fit a new cluster pipeline and compare if the predictions from this pipeline are "equivalent" to the predictions from the previous pipeline.*
* *Fit a classifier to explain cluster and check if performance on Train and Test sets is similar to the previous pipeline.*
* *Check if the most important features for the classifier are the same from the previous pipeline.*
* *Compare if the cluster profile from both pipelines is "equivalent".*

**To conclude** -

* Following a comprehensive comparative analysis of the existing and newly developed pipelines, I have determined that the newer pipeline, which incorporates only four features identified as optimal for cluster analysis, is more advantageous.

* This model retains diagnostic accuracy while enhancing computational efficiency.

* The streamlined approach ensures a clear distinction between benign and malignant clusters (indicated as 1 and 2), and it improves the detection of malignancies in borderline cases (designated as Cluster 0).

* Furthermore, the dimensionality reduction enhances its scalability for clinical implementation without compromising essential diagnostic performance.

## Conclusion About Clusters

* **Cluster 0:** Mostly malignant (69%), with moderately irregular contours (concavity 0.112-0.198), mid-range tumor size (perimeter 97.692-119.7), and high structural complexity (fractal dimension 0.104-0.124). The 31% benign cases suggest some borderline tumors in this group.

* **Cluster 1:** Predominantly benign (92%), featuring smooth margins (concavity 0.02-0.059), compact size (perimeter 79.7-98.385), and low structural disorder (fractal dimension 0.069-0.083). The 8% malignant cases may represent early-stage or misclassified tumors.

* **Cluster 2:** Entirely malignant (100%), showing highly irregular shapes (concavity 0.125-0.219), large tumor spread (perimeter 140.35-170.15), and moderate structural chaos (fractal dimension 0.076-0.092). This cluster clearly represents aggressive, high-risk malignancies.

pipeline_cluster # selected pipeline

## Push files to Repo

We will generate the following files

* Cluster Pipeline
* Train Set
* Feature importance plot
* Clusters Description
* Cluster Silhouette

import joblib
import os

version = 'v1'
file_path = f'outputs/ml_pipeline/cluster_analysis/{version}'

try:
    os.makedirs(name=file_path)
except Exception as e:
    print(e)

## Cluster pipeline

pipeline_cluster

joblib.dump(value=pipeline_cluster, filename=f"{file_path}/cluster_pipeline.pkl")

## Train Set

print(df_reduced.shape)
df_reduced.head(3)

df_reduced.to_csv(f"{file_path}/TrainSet.csv", index=False)

## Most important features plot

* These are the features that define a cluster

df_feature_importance.plot(kind='bar', x='Feature', y='Importance', figsize=(8,4))
plt.show()

df_feature_importance.plot(kind='bar', x='Feature', y='Importance', figsize=(8,4))
plt.savefig(f"{file_path}/features_define_cluster.png", bbox_inches='tight', dpi=150)

## Cluster Profile

clusters_profile

clusters_profile.to_csv(f"{file_path}/clusters_profile.csv")

## Cluster silhouette plot

visualizer = SilhouetteVisualizer(Pipeline(pipeline_cluster.steps[-1:])[0], colors='yellowbrick')
visualizer.fit(df_analysis)
visualizer.show()
plt.show()

fig, axes = plt.subplots(figsize=(7,5))
fig = SilhouetteVisualizer(Pipeline(pipeline_cluster.steps[-1:])[0] , colors='yellowbrick', ax=axes)
fig.fit(df_analysis)

plt.savefig(f"{file_path}/clusters_silhouette.png", bbox_inches='tight',dpi=150)

---
